view
view 파일 작성 및 라우팅 경로 재설정
