+++ templates/index.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    
+    <a href="/user">login</a>
 </body>
 </html>
